+++ cli/tests/test_config_command.rs
@@ -20,10 +20,22 @@ use itertools::Itertools as _;
 use regex::Regex;
 
 use crate::common::TestEnvironment;
+use crate::common::TestWorkDir;
 use crate::common::default_config_from_schema;
 use crate::common::fake_editor_path;
 use crate::common::force_interactive;
 use crate::common::to_toml_value;
+
+// Gets the directory for a given config id.
+fn config_dir<'a>(test_env: &'a TestEnvironment, dir: &str) -> TestWorkDir<'a> {
+    let d = test_env.work_dir("home/.config/jj").dir(dir);
+    // Find the config ID that was used.
+    let mut entries = std::fs::read_dir(d.root()).unwrap();
+    let subdir = entries.next().unwrap().unwrap().file_name();
+    // Doesn't handle multiple config IDs.
+    assert!(entries.next().is_none());
+    d.dir(subdir)
+}
 
 #[test]
 fn test_config_list_single() {
@@ -337,9 +349,9 @@ fn test_config_list_origin() {
         "--config",
         "test-cli-key=test-cli-val",
     ]);
-    insta::assert_snapshot!(output, @r#"
+    insta::assert_snapshot!(output, @r###"
     test-key = "test-val" # user $TEST_ENV/config/config.toml
-    test-layered-key = "test-layered-val" # repo $TEST_ENV/repo/.jj/repo/config.toml
+    test-layered-key = "test-layered-val" # repo $TEST_ENV/home/.config/jj/repos/$CONFIG_ID/config.toml
     user.name = "Test User" # env
     user.email = "test.user@example.com" # env
     debug.commit-timestamp = "2001-02-03T04:05:11+07:00" # env
@@ -349,7 +361,7 @@ fn test_config_list_origin() {
     operation.username = "test-username" # env
     test-cli-key = "test-cli-val" # cli
     [EOF]
-    "#);
+    "###);
 
     let output = work_dir.run_jj([
         "config",
@@ -724,7 +736,7 @@ fn test_config_set_for_repo() {
         .run_jj(["config", "set", "--repo", "test-table.foo", "true"])
         .success();
     // Ensure test-key successfully written to user config.
-    let repo_config_toml = work_dir.read_file(".jj/repo/config.toml");
+    let repo_config_toml = config_dir(&test_env, "repos").read_file("config.toml");
     insta::assert_snapshot!(repo_config_toml, @r#"
     #:schema https://docs.jj-vcs.dev/latest/config-schema.json
 
@@ -752,7 +764,7 @@ fn test_config_set_for_workspace() {
         .success();
 
     // Read workspace config
-    let workspace_config = work_dir.read_file(".jj/workspace-config.toml");
+    let workspace_config = config_dir(&test_env, "workspaces").read_file("config.toml");
     insta::assert_snapshot!(workspace_config, @r#"
     #:schema https://docs.jj-vcs.dev/latest/config-schema.json
 
@@ -968,7 +980,7 @@ fn test_config_unset_for_repo() {
         .run_jj(["config", "unset", "--repo", "test-key"])
         .success();
 
-    let repo_config_toml = work_dir.read_file(".jj/repo/config.toml");
+    let repo_config_toml = config_dir(&test_env, "repos").read_file("config.toml");
     insta::assert_snapshot!(repo_config_toml, @"");
 }
 
@@ -991,7 +1003,7 @@ fn test_config_unset_for_workspace() {
         .run_jj(["config", "unset", "--workspace", "foo"])
         .success();
 
-    let workspace_config = work_dir.read_file(".jj/workspace-config.toml");
+    let workspace_config = config_dir(&test_env, "workspaces").read_file("config.toml");
     insta::assert_snapshot!(workspace_config, @"");
 }
 
@@ -1055,17 +1067,18 @@ fn test_config_edit_repo() {
     let mut test_env = TestEnvironment::default();
     let edit_script = test_env.set_up_fake_editor();
     test_env.run_jj_in(".", ["git", "init", "repo"]).success();
-    let work_dir = test_env.work_dir("repo");
-    let repo_config_path = work_dir
-        .root()
-        .join(PathBuf::from_iter([".jj", "repo", "config.toml"]));
-    assert!(!repo_config_path.exists());
+    let repo_config_dir = test_env.home_dir().join(".config/jj/repos");
+    assert!(!repo_config_dir.is_dir());
 
     std::fs::write(edit_script, "dump-path path").unwrap();
+    let work_dir = test_env.work_dir("repo");
     work_dir.run_jj(["config", "edit", "--repo"]).success();
+
+    let repo_config_path = config_dir(&test_env, "repos").root().join("config.toml");
 
     let edited_path =
         PathBuf::from(std::fs::read_to_string(test_env.env_root().join("path")).unwrap());
+    assert!(repo_config_dir.is_dir());
     assert_eq!(edited_path, dunce::simplified(&repo_config_path));
     assert!(repo_config_path.exists(), "new file should be created");
 }
@@ -1088,9 +1101,9 @@ fn test_config_edit_invalid_config() {
             .args(["config", "edit", "--repo"])
             .write_stdin("Y\n")
     });
-    insta::assert_snapshot!(output, @r"
+    insta::assert_snapshot!(output, @r###"
     ------- stderr -------
-    Editing file: $TEST_ENV/repo/.jj/repo/config.toml
+    Editing file: $TEST_ENV/home/.config/jj/repos/$CONFIG_ID/config.toml
     Warning: An error has been found inside the config:
     Caused by:
     1: Configuration cannot be parsed as TOML document
@@ -1101,7 +1114,7 @@ fn test_config_edit_invalid_config() {
     key with no value, expected `=`
 
     Do you want to keep editing the file? If not, previous config will be restored. (Yn): [EOF]
-    ");
+    "###);
 
     let output = work_dir.run_jj(["config", "get", "test"]);
     insta::assert_snapshot!(output, @r"
@@ -1117,9 +1130,9 @@ fn test_config_edit_invalid_config() {
             .args(["config", "edit", "--repo"])
             .write_stdin("n\n")
     });
-    insta::assert_snapshot!(output, @r"
+    insta::assert_snapshot!(output, @r###"
     ------- stderr -------
-    Editing file: $TEST_ENV/repo/.jj/repo/config.toml
+    Editing file: $TEST_ENV/home/.config/jj/repos/$CONFIG_ID/config.toml
     Warning: An error has been found inside the config:
     Caused by:
     1: Configuration cannot be parsed as TOML document
@@ -1130,7 +1143,7 @@ fn test_config_edit_invalid_config() {
     key with no value, expected `=`
 
     Do you want to keep editing the file? If not, previous config will be restored. (Yn): [EOF]
-    ");
+    "###);
 
     let output = work_dir.run_jj(["config", "get", "test"]);
     insta::assert_snapshot!(output, @r"
@@ -1164,10 +1177,10 @@ fn test_config_path() {
         "jj config path shouldn't create new file"
     );
 
-    insta::assert_snapshot!(work_dir.run_jj(["config", "path", "--repo"]), @r"
-    $TEST_ENV/repo/.jj/repo/config.toml
+    insta::assert_snapshot!(work_dir.run_jj(["config", "path", "--repo"]), @r###"
+    $TEST_ENV/home/.config/jj/repos/$CONFIG_ID/config.toml
     [EOF]
-    ");
+    "###);
     assert!(
         !repo_config_path.exists(),
         "jj config path shouldn't create new file"